Introduce logic for building only one devnet or mainnet docker and debian package

+++ scripts/rebuild-deb.sh
@@ -129,14 +129,15 @@ ls -lh mina*.deb
 ##################################### END LOGPROC PACKAGE #######################################
 
 ##################################### GENERATE MINA MAINNET PACKAGE #######################################
+if [[ "${DUNE_PROFILE}" -eq "mainnet" ]]; then 
 
-###### deb without the proving keys
-echo "------------------------------------------------------------"
-echo "Building mainnet deb without keys:"
+  ###### deb without the proving keys
+  echo "------------------------------------------------------------"
+  echo "Building mainnet deb without keys:"
 
-rm -rf "${BUILDDIR}"
-mkdir -p "${BUILDDIR}/DEBIAN"
-cat << EOF > "${BUILDDIR}/DEBIAN/control"
+  rm -rf "${BUILDDIR}"
+  mkdir -p "${BUILDDIR}/DEBIAN"
+  cat << EOF > "${BUILDDIR}/DEBIAN/control"
 Package: mina-mainnet
 Version: ${MINA_DEB_VERSION}
 Section: base
@@ -153,77 +154,77 @@ Description: Mina Client and Daemon
  Built from ${GITHASH} by ${BUILD_URL}
 EOF
 
-echo "------------------------------------------------------------"
-echo "Control File:"
-cat "${BUILDDIR}/DEBIAN/control"
+  echo "------------------------------------------------------------"
+  echo "Control File:"
+  cat "${BUILDDIR}/DEBIAN/control"
 
-echo "------------------------------------------------------------"
-# Binaries
-mkdir -p "${BUILDDIR}/usr/local/bin"
-sudo cp ./default/src/app/cli/src/mina_mainnet_signatures.exe "${BUILDDIR}/usr/local/bin/mina"
-sudo cp ./default/src/app/rosetta/rosetta_mainnet_signatures.exe "${BUILDDIR}/usr/local/bin/mina-rosetta"
+  echo "------------------------------------------------------------"
+  # Binaries
+  mkdir -p "${BUILDDIR}/usr/local/bin"
+  sudo cp ./default/src/app/cli/src/mina_mainnet_signatures.exe "${BUILDDIR}/usr/local/bin/mina"
+  sudo cp ./default/src/app/rosetta/rosetta_mainnet_signatures.exe "${BUILDDIR}/usr/local/bin/mina-rosetta"
 
-libp2p_location=../src/app/libp2p_helper/result/bin
-ls -l ../src/app/libp2p_helper/result/bin || libp2p_location=$HOME/app/
-p2p_path="${BUILDDIR}/usr/local/bin/coda-libp2p_helper"
-cp $libp2p_location/libp2p_helper $p2p_path
-chmod +w $p2p_path
-# Only for nix builds
-# patchelf --set-interpreter /lib64/ld-linux-x86-64.so.2 "${BUILDDIR}/usr/local/bin/coda-libp2p_helper"
-chmod -w $p2p_path
-# cp ./default/src/app/logproc/logproc.exe "${BUILDDIR}/usr/local/bin/mina-logproc"
-cp ./default/src/app/runtime_genesis_ledger/runtime_genesis_ledger.exe "${BUILDDIR}/usr/local/bin/mina-create-genesis"
+  libp2p_location=../src/app/libp2p_helper/result/bin
+  ls -l ../src/app/libp2p_helper/result/bin || libp2p_location=$HOME/app/
+  p2p_path="${BUILDDIR}/usr/local/bin/coda-libp2p_helper"
+  cp $libp2p_location/libp2p_helper $p2p_path
+  chmod +w $p2p_path
+  # Only for nix builds
+  # patchelf --set-interpreter /lib64/ld-linux-x86-64.so.2 "${BUILDDIR}/usr/local/bin/coda-libp2p_helper"
+  chmod -w $p2p_path
+  # cp ./default/src/app/logproc/logproc.exe "${BUILDDIR}/usr/local/bin/mina-logproc"
+  cp ./default/src/app/runtime_genesis_ledger/runtime_genesis_ledger.exe "${BUILDDIR}/usr/local/bin/mina-create-genesis"
 
-mkdir -p "${BUILDDIR}/usr/lib/systemd/user"
-sed s%PEERS_LIST_URL_PLACEHOLDER%https://storage.googleapis.com/mina-seed-lists/mainnet_seeds.txt% ../scripts/mina.service > "${BUILDDIR}/usr/lib/systemd/user/mina.service"
+  mkdir -p "${BUILDDIR}/usr/lib/systemd/user"
+  sed s%PEERS_LIST_URL_PLACEHOLDER%https://storage.googleapis.com/mina-seed-lists/mainnet_seeds.txt% ../scripts/mina.service > "${BUILDDIR}/usr/lib/systemd/user/mina.service"
 
-# Build Config
-mkdir -p "${BUILDDIR}/etc/coda/build_config"
-cp ../src/config/mainnet.mlh "${BUILDDIR}/etc/coda/build_config/BUILD.mlh"
-rsync -Huav ../src/config/* "${BUILDDIR}/etc/coda/build_config/."
+  # Build Config
+  mkdir -p "${BUILDDIR}/etc/coda/build_config"
+  cp ../src/config/mainnet.mlh "${BUILDDIR}/etc/coda/build_config/BUILD.mlh"
+  rsync -Huav ../src/config/* "${BUILDDIR}/etc/coda/build_config/."
 
-# Copy the genesis ledgers and proofs as these are fairly small and very valueable to have l
-# Genesis Ledger/proof/epoch ledger Copy
-mkdir -p "${BUILDDIR}/var/lib/coda"
-for f in /tmp/coda_cache_dir/genesis*; do
-    if [ -e "$f" ]; then
-        mv /tmp/coda_cache_dir/genesis* "${BUILDDIR}/var/lib/coda/."
-    fi
-done
+  # Copy the genesis ledgers and proofs as these are fairly small and very valueable to have l
+  # Genesis Ledger/proof/epoch ledger Copy
+  mkdir -p "${BUILDDIR}/var/lib/coda"
+  for f in /tmp/coda_cache_dir/genesis*; do
+      if [ -e "$f" ]; then
+          mv /tmp/coda_cache_dir/genesis* "${BUILDDIR}/var/lib/coda/."
+      fi
+  done
 
-#copy config.json
-cp ../genesis_ledgers/mainnet.json "${BUILDDIR}/var/lib/coda/mainnet.json"
-cp ../genesis_ledgers/devnet.json "${BUILDDIR}/var/lib/coda/devnet.json"
-# The default configuration:
-cp ../genesis_ledgers/mainnet.json "${BUILDDIR}/var/lib/coda/config_${GITHASH_CONFIG}.json"
+  #copy config.json
+  cp ../genesis_ledgers/mainnet.json "${BUILDDIR}/var/lib/coda/mainnet.json"
+  cp ../genesis_ledgers/devnet.json "${BUILDDIR}/var/lib/coda/devnet.json"
+  # The default configuration:
+  cp ../genesis_ledgers/mainnet.json "${BUILDDIR}/var/lib/coda/config_${GITHASH_CONFIG}.json"
 
-# Bash autocompletion
-# NOTE: We do not list bash-completion as a required package,
-#       but it needs to be present for this to be effective
-mkdir -p "${BUILDDIR}/etc/bash_completion.d"
-cwd=$(pwd)
-export PATH=${cwd}/${BUILDDIR}/usr/local/bin/:${PATH}
-env COMMAND_OUTPUT_INSTALLATION_BASH=1 mina  > "${BUILDDIR}/etc/bash_completion.d/mina"
+  # Bash autocompletion
+  # NOTE: We do not list bash-completion as a required package,
+  #       but it needs to be present for this to be effective
+  mkdir -p "${BUILDDIR}/etc/bash_completion.d"
+  cwd=$(pwd)
+  export PATH=${cwd}/${BUILDDIR}/usr/local/bin/:${PATH}
+  env COMMAND_OUTPUT_INSTALLATION_BASH=1 mina  > "${BUILDDIR}/etc/bash_completion.d/mina"
 
-# echo contents of deb
-echo "------------------------------------------------------------"
-echo "Deb Contents:"
-find "${BUILDDIR}"
+  # echo contents of deb
+  echo "------------------------------------------------------------"
+  echo "Deb Contents:"
+  find "${BUILDDIR}"
 
-# Build the package
-echo "------------------------------------------------------------"
-fakeroot dpkg-deb --build "${BUILDDIR}" mina-mainnet_${MINA_DEB_VERSION}.deb
-ls -lh mina*.deb
-
+  # Build the package
+  echo "------------------------------------------------------------"
+  fakeroot dpkg-deb --build "${BUILDDIR}" mina-mainnet_${MINA_DEB_VERSION}.deb
+  ls -lh mina*.deb
+fi
 ##################################### END GENERATE MINA MAINNET PACKAGE #######################################
 
 ##################################### GENERATE MINA DEVNET PACKAGE #######################################
+if [[ "${DUNE_PROFILE}" -eq "devnet" ]]; then 
+  ###### deb with testnet signatures
+  echo "------------------------------------------------------------"
+  echo "Building testnet signatures deb without keys:"
 
-###### deb with testnet signatures
-echo "------------------------------------------------------------"
-echo "Building testnet signatures deb without keys:"
-
-cat << EOF > "${BUILDDIR}/DEBIAN/control"
+  cat << EOF > "${BUILDDIR}/DEBIAN/control"
 Package: mina-devnet
 Version: ${MINA_DEB_VERSION}
 Section: base
@@ -236,38 +237,39 @@ License: Apache-2.0
 Homepage: https://minaprotocol.com/
 Maintainer: o(1)Labs <build@o1labs.org>
 Description: Mina Client and Daemon
- Mina Protocol Client and Daemon
- Built from ${GITHASH} by ${BUILD_URL}
+Mina Protocol Client and Daemon
+Built from ${GITHASH} by ${BUILD_URL}
 EOF
 
-echo "------------------------------------------------------------"
-echo "Control File:"
-cat "${BUILDDIR}/DEBIAN/control"
+  echo "------------------------------------------------------------"
+  echo "Control File:"
+  cat "${BUILDDIR}/DEBIAN/control"
 
 
-echo "------------------------------------------------------------"
-# Overwrite binaries (sudo to fix permissions error)
-sudo cp ./default/src/app/cli/src/mina_testnet_signatures.exe "${BUILDDIR}/usr/local/bin/mina"
-sudo cp ./default/src/app/rosetta/rosetta_testnet_signatures.exe "${BUILDDIR}/usr/local/bin/mina-rosetta"
+  echo "------------------------------------------------------------"
+  # Overwrite binaries (sudo to fix permissions error)
+  sudo cp ./default/src/app/cli/src/mina_testnet_signatures.exe "${BUILDDIR}/usr/local/bin/mina"
+  sudo cp ./default/src/app/rosetta/rosetta_testnet_signatures.exe "${BUILDDIR}/usr/local/bin/mina-rosetta"
 
-# Switch the default configuration to devnet.json:
-sudo cp ../genesis_ledgers/devnet.json "${BUILDDIR}/var/lib/coda/config_${GITHASH_CONFIG}.json"
+  # Switch the default configuration to devnet.json:
+  sudo cp ../genesis_ledgers/devnet.json "${BUILDDIR}/var/lib/coda/config_${GITHASH_CONFIG}.json"
 
-# Overwrite the mina.service with a new default PEERS_URL
-rm -f "${BUILDDIR}/usr/lib/systemd/user/mina.service"
-sed s%PEERS_LIST_URL_PLACEHOLDER%https://storage.googleapis.com/seed-lists/devnet_seeds.txt% ../scripts/mina.service > "${BUILDDIR}/usr/lib/systemd/user/mina.service"
+  # Overwrite the mina.service with a new default PEERS_URL
+  rm -f "${BUILDDIR}/usr/lib/systemd/user/mina.service"
+  sed s%PEERS_LIST_URL_PLACEHOLDER%https://storage.googleapis.com/seed-lists/devnet_seeds.txt% ../scripts/mina.service > "${BUILDDIR}/usr/lib/systemd/user/mina.service"
 
 
-# echo contents of deb
-echo "------------------------------------------------------------"
-echo "Deb Contents:"
-find "${BUILDDIR}"
+  # echo contents of deb
+  echo "------------------------------------------------------------"
+  echo "Deb Contents:"
+  find "${BUILDDIR}"
 
-# Build the package
-echo "------------------------------------------------------------"
-fakeroot dpkg-deb --build "${BUILDDIR}" mina-devnet_${MINA_DEB_VERSION}.deb
-ls -lh mina*.deb
+  # Build the package
+  echo "------------------------------------------------------------"
+  fakeroot dpkg-deb --build "${BUILDDIR}" mina-devnet_${MINA_DEB_VERSION}.deb
+  ls -lh mina*.deb
 
+fi
 ##################################### END GENERATE MINA DEVNET PACKAGE #######################################
 
 ##################################### GENERATE MINA CREATE LEGACY GENESIS PACKAGE #######################################
